+++ packages/wallet-state/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './accounts'
 export * from './approval'
 export * from './base'
 export * from './discovery'