+++ packages/wallet-state/src/reducers/index.ts
@@ -1,0 +1,7 @@
+export { accountActions } from './accounts'
+export { discoveryActions } from './discovery'
+export { globalActions } from './global'
+export { keyringsActions } from './keyrings'
+export { settingsActions } from './settings'
+export { transactionsActions } from './transactions'
+export { uiActions } from './ui'