+++ packages/wallet-state/src/types/ui.ts
@@ -1,0 +1,30 @@
+export enum AssetTabKey {
+  ORDINALS = 0,
+  ATOMICALS = 1, // IGNORED
+  RUNES = 2,
+  CAT = 3,
+  ALKANES = 4,
+}
+
+export enum OrdinalsAssetTabKey {
+  ALL = 0,
+  BRC20 = 1,
+  BRC20_6BYTE = 2,
+}
+
+export enum CATAssetTabKey {
+  CAT20,
+  CAT721,
+  CAT20_V2,
+  CAT721_V2,
+}
+
+export enum AlkanesAssetTabKey {
+  TOKEN,
+  COLLECTION,
+}
+
+export enum NavigationSource {
+  BACK,
+  NORMAL,
+}