+++ packages/wallet-state/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/types/index.ts'],
   format: ['cjs', 'esm'],
   dts: true,
   clean: true,
@@ -10,5 +10,15 @@ export default defineConfig({
   minify: false,
   treeshake: true,
   outDir: 'lib',
-  external: ['react', 'react-redux', '@reduxjs/toolkit', '@unisat/wallet-types'],
+  external: [
+    'react',
+    'react-redux',
+    '@reduxjs/toolkit',
+    '@unisat/wallet-types',
+    '@unisat/wallet-shared',
+    '@unisat/babylon-service',
+    '@unisat/keyring-service',
+    '@unisat/base-utils',
+    '@unisat/i18n',
+  ],
 })