+++ packages/wallet-shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @unisat/wallet-shared
 
+## 0.1.2
+
+### Patch Changes
+
+- fix
+
 ## 0.1.1
 
 ### Patch Changes