+++ packages/wallet-state/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @unisat/wallet-state
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @unisat/base-utils@0.1.3
+
+## 1.0.4
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @unisat/wallet-shared@0.1.2
+
+## 1.0.3
+
+### Patch Changes
+
+- fix
+
 ## 1.0.2
 
 ### Patch Changes